refactor(server): forward telemetry to PostHog instead of self-hosting it
The Worker used to be the whole analytics stack: D1 for the daily budget
counters and the consent choices, Analytics Engine for the events, and an
/admin/query endpoint with a hand-rolled read-only SQL validator to get
anything back out of it. That is a lot of surface to maintain for five
numbers a month.

It is now a proxy. It stores nothing, holds no storage binding at all, and
forwards batches to PostHog EU Cloud. The privacy model is unchanged and is
the reason the proxy exists rather than the client talking to PostHog
directly: the Mode A daily hash is derived from the client IP, which the
client cannot see, and the personal API key that makes /forget and /export
possible cannot ship inside a desktop binary.

Every forwarded event overwrites $ip with a truthy placeholder and sets
$geoip_disable. The placeholder is not null on purpose: PostHog back-fills
$ip from the request socket whenever the submitted value is falsy, so null
would store an address rather than suppress one.

Mode A events carry $process_person_profile: false, so no person is created
and unique counting still works through the distinct_id.

The directory moved from telemetry/ to server/ because it is a deployable
Worker, not a description of what the app collects. The deployed Worker name
stays accshift-telemetry: the endpoint is compiled into binaries already in
the wild, and renaming would orphan them.

The D1 daily budget is gone with the database. The global rate limiter is a
burst ceiling keyed on a constant, which is a weaker guarantee; the hard
spend stop is now the PostHog project billing limit.
