--- telemetry/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "accshift-telemetry",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "wrangler dev",
-    "deploy": "wrangler d1 migrations apply accshift-telemetry --remote && wrangler deploy",
-    "deploy:worker-only": "wrangler deploy",
-    "db:migrate": "wrangler d1 migrations apply accshift-telemetry --remote",
-    "db:migrate:local": "wrangler d1 migrations apply accshift-telemetry --local",
-    "db:status": "wrangler d1 migrations list accshift-telemetry --remote",
-    "tail": "wrangler tail"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20260201.0",
-    "typescript": "^5.7.2",
-    "wrangler": "^4.84.1"
-  }
-}
